Player movement and mob placement works.
