--- server/game_components/entities/mob.py
@@ -1,5 +1,0 @@
-from entity import Entity
-
-
-class Mob(Entity):
-    pass
