--- server/game_components/entities/player.py
@@ -1,6 +1,0 @@
-from entity import Entity
-
-
-class Player(Entity):
-    experience: int
-    level: int
